+++ bin/commands/penaltyboxconditions.delete.js
@@ -1,0 +1,43 @@
+let PenaltyBoxConditions = require('../../src/penaltyboxconditions').penaltyBoxConditions;
+let out = require('./lib/out');
+
+class DeletePenaltyBoxConditionsCommand {
+  constructor() {
+    this.flags = 'delete-penalty-box-conditions';
+    this.desc = 'Delete penalty box conditions in a policy.';
+    this.setup = this.setup.bind(this);
+    this.run = this.run.bind(this);
+  }
+
+  setup(sywac) {
+    sywac
+      .number('--config <id>', {
+        desc: 'Configuration ID. Mandatory if you have more than one configuration.',
+        group: 'Optional:',
+        required: false
+      })
+      .string('--version <id>', {
+        desc:
+          "Version Number. It can also take the values 'PROD' or 'PRODUCTION' or 'STAGING'. If not provided, latest version is assumed.",
+        group: 'Optional:',
+        required: false
+      })
+      .string('--policy <id>', {
+        desc:
+          'Policy ID. If not provided, we try to use the policy available on file. If you have more than one policy, this option must be provided.',
+        group: 'Optional:',
+        required: false
+      });
+  }
+  run(options) {
+    out.print({
+      promise: new PenaltyBoxConditions(options).deletePenaltyBoxConditions(),
+      args: options,
+      success: (args, data) => {
+        return JSON.stringify(data);
+      }
+    });
+  }
+}
+
+module.exports = new DeletePenaltyBoxConditionsCommand();